pyproject.toml
[project]
name = "receiptly-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Vedran Selak"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13"
fastapi = "^0.115.12"
ruff = "^0.11.7"
pydantic-settings = "^2.9.1"
asyncpg = "^0.30.0"
alembic = "^1.15.2"
uvicorn = "^0.34.2"
passlib = "^1.7.4"
email-validator = "^2.2.0"
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
pyjwt = "^2.10.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
